+++ source/renderer/app/components/loading/Loading.js
@@ -94,6 +94,11 @@ const messages = defineMessages({
     defaultMessage: '!!!Download logs',
     description: 'Download logs button label on the loading.',
   },
+  tlsCertificateNotValidError: {
+    id: 'loading.screen.errors.tlsCertificateNotValidPleaseRestartError',
+    defaultMessage: '!!!TLS certificate is not valid, please restart Daedalus.',
+    description: 'The TLS cert is not valid and Daedalus should be restarted',
+  },
 });
 
 type State = {
@@ -112,6 +117,7 @@ type Props = {
   isNodeStopping: boolean,
   isNodeStopped: boolean,
   isNotEnoughDiskSpace: boolean,
+  isTlsCertInvalid: boolean,
   diskSpaceRequired: string,
   diskSpaceMissing: string,
   diskSpaceRecommended: string,
@@ -239,7 +245,7 @@ export default class Loading extends Component<Props, State> {
   };
 
   _getConnectingMessage = () => {
-    const { cardanoNodeState, hasBeenConnected } = this.props;
+    const { cardanoNodeState, hasBeenConnected, isTlsCertInvalid } = this.props;
     let connectingMessage;
     switch (cardanoNodeState) {
       case null:
@@ -272,6 +278,12 @@ export default class Loading extends Component<Props, State> {
           ? messages.reconnecting
           : messages.connecting;
     }
+    const isConnectingMessage =
+      connectingMessage === messages.connecting ||
+      connectingMessage === messages.reconnecting;
+    if (isTlsCertInvalid && isConnectingMessage) {
+      return messages.tlsCertificateNotValidError;
+    }
     return connectingMessage;
   };
 
@@ -284,6 +296,7 @@ export default class Loading extends Component<Props, State> {
       isNodeStopping,
       isNodeStopped,
       isNotEnoughDiskSpace,
+      isTlsCertInvalid,
       diskSpaceRequired,
       diskSpaceMissing,
       diskSpaceRecommended,
@@ -319,11 +332,12 @@ export default class Loading extends Component<Props, State> {
         />
       );
     }
+    const showEllipsis = isNodeStopped || (isTlsCertInvalid && !isNodeStopping);
 
     if (!isConnected) {
       const headlineClasses = classNames([
         styles.headline,
-        isNodeStopped ? styles.withoutAnimation : null,
+        showEllipsis ? styles.withoutAnimation : null,
       ]);
       return (
         <div className={styles.connecting}>